requirements.txt
annotated-types==0.7.0
anyio==4.13.0
audioop-lts==0.2.2
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.1.8
colorama==0.4.6
comtypes==1.4.16
cryptography==46.0.7
distro==1.9.0
docopt==0.6.2
et_xmlfile==2.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.194.0
google-auth==2.49.1
google-auth-httplib2==0.3.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
groq==1.1.2
grpcio==1.80.0
grpcio-status==1.71.2
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
Js2Py==0.74
lxml==6.0.2
MouseInfo==0.1.3
numpy==2.4.4
openpyxl==3.1.5
packaging==26.0
pandas==3.0.2
pillow==12.2.0
pipwin==0.5.2
proto-plus==1.27.2
protobuf==5.29.6
psutil==7.2.2
pyasn1==0.6.3
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
PyGetWindow==0.0.9
pyjsparser==2.7.1
PyMsgBox==2.0.1
pyparsing==3.3.2
PyPDF2==3.0.1
pyperclip==1.11.0
pypiwin32==223
PyPrind==2.11.3
PyRect==0.2.0
PyScreeze==1.0.1
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-pptx==1.0.2
python-telegram-bot==22.7
pyttsx3==2.99
pytweening==1.2.0
pywin32==311
requests==2.33.1
scipy==1.17.1
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soundfile==0.13.1
soupsieve==2.8.3
SpeechRecognition==3.16.0
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
xlsxwriter==3.2.9
